break out commands, git functions, types, init function; add some comments
