+++ cmd/net-helper/init.go
@@ -1,0 +1,40 @@
+package main
+
+import (
+	"fmt"
+	"io"
+	"io/ioutil"
+
+	cli "github.com/urfave/cli/v2"
+)
+
+/*
+	InitApp()
+
+	- set multiple templates and printers for cli app
+	- simply a way to customize the underlying behavior
+*/
+func InitApp() {
+	cli.AppHelpTemplate += "\nThis might be helpful: \n"
+	cli.CommandHelpTemplate += "\nThis is command help: \n"
+	cli.SubcommandHelpTemplate += "\nThis is subcommand help"
+
+	cli.HelpFlag = &cli.BoolFlag{Name: "help", Aliases: []string{"h", "H", "halp", "wtf"}}
+	cli.VersionFlag = &cli.BoolFlag{Name: "print-version", Aliases: []string{"V", "v", "ver"}}
+
+	cli.HelpPrinter = func(w io.Writer, templ string, data interface{}) {
+		fmt.Fprintf(w, "Never fear, HELP is here: ")
+	}
+	cli.VersionPrinter = func(c *cli.Context) {
+		fmt.Fprintf(c.App.Writer, "version=%s\n", c.App.Version)
+	}
+	cli.OsExiter = func(c int) {
+		fmt.Fprintf(cli.ErrWriter, "refusing to exit %d\n", c)
+	}
+
+	cli.ErrWriter = ioutil.Discard
+
+	cli.FlagStringer = func(fl cli.Flag) string {
+		return fmt.Sprintf("\t\t%s", fl.Names()[0])
+	}
+}
